Small bug fixes. Slightly changed the GUI and its logic.
Now, the sending worker process can be paused and
resumed. The entire process can also be stopped
and started again. The user can see the status of
the worker process.

Fixed a bug in the to_bytes() method of the
HelloPacket and the HalloReplyPacket classes.

Other small changes.

There are still problems when changing the network
interface in the GUI. This will be addressed soon.