Fixed an issue with displaying the log messages in the GUI.
The log messages dialog window was updated in the
logger thread which led to sudden crashes. The
logging handler was updated to use Qt's signal and
slot mechanism to notify the GUI thread when new
log messages are available (queued connection /
defered messages). This ensures that the GUI
thread is always in control of the GUI.